requirements.txt
django==2.2
pillow
mysqlclient
djangorestframework
markdown
django-filter
django-cors-headers
django-guardian
coreapi
Pygments
djangorestframework-jwt
requests